Fix security holes and wire up the Discord bot for deployment
- Remove unauthenticated /api/discord/activity/ webhook that let anyone
  inflate any user's reputation by guessing their discord_id
- Add OAuth state param to Discord connect/callback to prevent
  account-linking CSRF
- Remove dev-only script that created a hardcoded admin/admin123 superuser
- Add worker process to Procfile so the Discord chat bot (run_chat_bot)
  actually runs instead of never being started
- Remove broken run_welcome_bot command (imported a function that
  doesn't exist in discord_bot.py)
- Change discord_chat_points to FloatField so fractional points from
  chat sessions stop getting silently truncated
- Remove dead/unused code: broken serializers.py (referenced a deleted
  Referral model), no-op HealthcheckBypassMiddleware, unrouted
  healthcheck files, unused start.sh/healthcheck_server.py/test_server.py
- Strip ~13 unused Solana/crypto packages from requirements.txt
- Strip verbose emoji debug prints from settings.py/wsgi.py/__init__.py
  (also crashed under non-UTF8 locales)
- Fix .gitignore: was ignoring static/ (tracked source assets) instead
  of staticfiles/ (collectstatic output); add *.log
- Add .env.example documenting required env vars

+++ crownie/models.py
@@ -18,7 +18,7 @@ class User(AbstractUser):
     discord_connected_at = models.DateTimeField(blank=True, null=True)
     
     # Discord chat tracking fields
-    discord_chat_points = models.IntegerField(default=0)
+    discord_chat_points = models.FloatField(default=0)
     discord_total_messages = models.IntegerField(default=0)
     discord_last_activity = models.DateTimeField(blank=True, null=True)
